Wait for GHA eval results instead of falling back to local evaluation

+++ nixpkgs_review/builddir.py
@@ -85,10 +85,11 @@ class Builddir:
         os.environ.update(self.environ)
 
         with DisableKeyboardInterrupt():
-            res = sh(["git", "worktree", "remove", "-f", str(self.worktree_dir)])
-            if res.returncode != 0:
-                warn(
-                    f"Failed to remove worktree at {self.worktree_dir}. Please remove it manually. Git failed with: {res.returncode}"
-                )
+            if Path.exists(self.worktree_dir / ".git"):
+                res = sh(["git", "worktree", "remove", "-f", str(self.worktree_dir)])
+                if res.returncode != 0:
+                    warn(
+                        f"Failed to remove worktree at {self.worktree_dir}. Please remove it manually. Git failed with: {res.returncode}"
+                    )
 
         self.overlay.cleanup()
